Results: refactor api/watcher packages into separate pkg subfolders.
This moves packages into pkg/api or pkg/watcher subfolders based on
which component they belong to. This also breaks out some logic defined
in cmd/watcher into its own pkg/watcher/reconciler package (name chosen
to ~match pipeline naming). This does not introduce any new behavior.
